### 2017-11-14 #455 @whedon
[PRE REVIEW]: A framework for generating interactive reports for cancer genome analysis
**Submitting author:** @aokad (<a href="http://orcid.org/0000-0002-9782-5900">Ai Okada</a>)
**Repository:** <a href="https://github.com/Genomon-Project/paplot/" target ="_blank">https://github.com/Genomon-Project/paplot/</a>
**Version:** 0.5.5
