### 2017-11-20 #457 @aokad
Dear @pjotrp, @chfi,

I really appreciate the prompt response to our submission.
As far as I confirmed, I believe our software satisfies most of the check list except:

1. Automated tests: Our repository currently supports automated test for the build. We will further improve so that several command executions can be checked soon.
2. Community guidelines: We will soon include clearer statement in the document.
3. References: We will improve the reference list so that DOIs are included.
