+++ paper/paper.bib
@@ -1,6 +1,7 @@
 @article{alexandrov2013deciphering,
   title={Deciphering signatures of mutational processes operative in human cancer},
   author={Alexandrov, Ludmil B and Nik-Zainal, Serena and Wedge, David C and Campbell, Peter J and Stratton, Michael R},
+  doi = {10.1016/j.celrep.2012.12.008},
   journal={Cell reports},
   volume={3},
   number={1},
@@ -12,6 +13,7 @@
 @article{kataoka2016aberrant,
   title={Aberrant PD-L1 expression through 3′-UTR disruption in multiple cancers},
   author={Kataoka, Keisuke and Shiraishi, Yuichi and Takeda, Yohei and Sakata, Seiji and Matsumoto, Misako and Nagano, Seiji and Maeda, Takuya and Nagata, Yasunobu and Kitanaka, Akira and Mizuno, Seiya and others},
+  doi = {10.1038/nature18294},
   journal={Nature},
   volume={534},
   number={7607},
@@ -23,6 +25,7 @@
 @article{koboldt2012varscan,
   title={VarScan 2: somatic mutation and copy number alteration discovery in cancer by exome sequencing},
   author={Koboldt, Daniel C and Zhang, Qunyuan and Larson, David E and Shen, Dong and McLellan, Michael D and Lin, Ling and Miller, Christopher A and Mardis, Elaine R and Ding, Li and Wilson, Richard K},
+  doi = {10.1101/gr.129684.111},
   journal={Genome research},
   volume={22},
   number={3},
@@ -34,6 +37,7 @@
 @article{krzywinski2009circos,
   title={Circos: an information aesthetic for comparative genomics},
   author={Krzywinski, Martin and Schein, Jacqueline and Birol, Inanc and Connors, Joseph and Gascoyne, Randy and Horsman, Doug and Jones, Steven J and Marra, Marco A},
+  doi = {10.1101/gr.092759.109},
   journal={Genome research},
   volume={19},
   number={9},
@@ -45,6 +49,7 @@
 @article{mitelman2007impact,
   title={The impact of translocations and gene fusions on cancer causation},
   author={Mitelman, Felix and Johansson, Bertil and Mertens, Fredrik},
+  doi = {10.1038/nrc2091},
   journal={Nature Reviews Cancer},
   volume={7},
   number={4},
@@ -56,6 +61,7 @@
 @article{rausch2012delly,
   title={DELLY: structural variant discovery by integrated paired-end and split-read analysis},
   author={Rausch, Tobias and Zichner, Thomas and Schlattl, Andreas and St{\"u}tz, Adrian M and Benes, Vladimir and Korbel, Jan O},
+  doi = {10.1093/bioinformatics/bts378},
   journal={Bioinformatics},
   volume={28},
   number={18},
@@ -67,6 +73,7 @@
 @article{shiraishi2013empirical,
   title={An empirical Bayesian framework for somatic mutation detection from cancer genome sequencing data},
   author={Shiraishi, Yuichi and Sato, Yusuke and Chiba, Kenichi and Okuno, Yusuke and Nagata, Yasunobu and Yoshida, Kenichi and Shiba, Norio and Hayashi, Yasuhide and Kume, Haruki and Homma, Yukio and others},
+  doi = {10.1093/nar/gkt126},
   journal={Nucleic acids research},
   volume={41},
   number={7},
@@ -78,6 +85,7 @@
 @article{shiraishi2015simple,
   title={A simple model-based approach to inferring and visualizing cancer mutation signatures},
   author={Shiraishi, Yuichi and Tremmel, Georg and Miyano, Satoru and Stephens, Matthew},
[…]
