### 2017-11-21 b78667a
remove debug print
