+++ tests/ref/ca/test1_01_sample_header_id_csv/data_ca.js
@@ -37,9 +37,9 @@ ca_data.tooltip_format = { bundle:{format:[[{label:'[',type:'fix',keys:[],ext:''
 ca_data.link_header = ['dir1','dir2','gene_name1','gene_name2','type',];
 // 0:ID, 1:chr1, 2:break1, 3:chr2, 4:break2, 5:is_outer, 6:group_id, 7:tooltip_data
 ca_data.links = [["SAMPLE00","13",16019088,"11",62784483,false,0,[['-','+','LS7T1EG444','4GRRIO5AVR','deletion',],]],["SAMPLE00","08",99412502,"06",129302434,false,0,[['-','+','FQFW16UF5U','QP779MLPNV','translocation',],]],["SAMPLE00","12",84663781,"17",52991509,false,0,[['+','-','Q9VX1I9U3I','7XM09ETN40','deletion',],]],["SAMPLE00","00",153160367,"21",33751554,false,0,[['+','+','CEE2SPV1R1','PVYYQIVS8G','inversion',],]],["SAMPLE00","17",12249358,"02",146222593,false,0,[['-','+','HH9OL7CK6G','XD80LI4E6Q','translocation',],]],["SAMPLE00","20",8658030,"22",133492043,false,0,[['+','-','I20EVP15ZM','WPE8O5H237','tandem_duplication',],]],["SAMPLE00","11",120178477,"00",155354923,false,0,[['+','-','IMYXD3TCA4','3MNN5J0MDN','deletion',],]],["SAMPLE01","10",101374238,"21",26701405,false,0,[['+','+','FZ7LOS66RD','9WYBJR57E0','translocation',],]],["SAMPLE01","01",121708638,"06",137424167,false,0,[['-','-','5655M5E46B','HB14VJXDHV','translocation',],]],["SAMPLE01","15",43027789,"21",23791492,false,0,[['+','-','REFSIL0H2M','L5EA31R8U0','inversion',],]],["SAMPLE01","18",3862589,"15",37135239,false,0,[['-','+','1IRWHVZLH8','6FUR9YMZOH','deletion',],]],["SAMPLE01","19",50294222,"00",164250235,false,0,[['+','-','DOH5G0YRQ9','9TWYMR5CZ2','inversion',],]],["SAMPLE01","22",67392415,"14",3327412,false,0,[['+','+','EM36MRX9B3','G4FPLN527D','translocation',],]],["SAMPLE02","21",34268355,"09",19871820,false,0,[['+','+','9SVRQCFVCO','2BEWSO91FZ','tandem_duplication',],]],["SAMPLE02","01",51849987,"19",47304716,false,0,[['-','+','9HK7LTVZST','DNCS44WCW0','inversion',],]],["SAMPLE02","05",66742149,"05",31916435,true,1,[['-','+','TJJCWQ6K6J','V5X8Y5QHVS','translocation',],]],];
-ca_data.select_value = [[1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1],[1,1],];
-ca_data.select_key = [[[15,7],[18,0],[15,8],[17,10],[0,30],[0,31],[0,32],[19,10],[21,5],[21,4],[1,10],[10,20],[20,1],[6,25],[6,27],[11,24],[12,16],[22,26],[9,3],[2,29],[17,2],[21,6],[8,19],[1,24],[11,12],[13,3],[19,9],[14,0],[22,13]],[[5,6],[5,13]],];
[…]
