+++ tests/unit_tests.py
@@ -24,5 +24,5 @@ def suite():
     
     import test_pmsignature
     suite.addTests(unittest.makeSuite(test_pmsignature.TestSet))
-
+    
     return suite
