### 2017-11-22 #457 @chfi
Dear @aokad, your submission looks very good. Installing and running was a breeze, and everything appears to work just as the excellent documentation says it should. The only things I found missing are the three parts you already identified and are working on :)
