### 2017-11-24 #457 @aokad
Dear @pjotrp, @chfi,
I have resolved.

1. Automated tests: I added "python setup.py test", which checks the validity of several sub-commands for generating analytical reports.
2. Community guidelines: I have included some description on community guideline to the Contact section in the documentation (http://paplot-doc.readthedocs.io/en/latest/info.html#contact)
3. References: I added DOIs to paper/paper.bib.

Please kindly let me.
