### 2017-11-27 #457 @arfon
@aokad - apologies for the late reply to this thread but could you please add a little more detail to your paper. Please see the guidance for authors [here](http://joss.theoj.org/about#paper_structure).
