### 2017-11-27 #457 @pjotrp
Yes, main thing is to make it clear to readers what it is about. See it as an advertisement for your work. A figure is usually a good idea. See one of our own [papers](http://joss.theoj.org/papers/84b27b28948be80a47268a5baa2701ad) as an example.
