+++ paper/paper.md
@@ -26,6 +26,56 @@ bibliography: paper.bib
 
 # Summary
 
-We introduce paplot, the software for generating dynamic reports that are frequently necessary in the post analyticalphases of cancer genome studies. 
-The "interactive" nature of the paplot-generated reportsenables users to extract much richer information than that obtained from static graphs via most conventional visualization tools.
+We introduce paplot, the software for generating dynamic reports that are frequently necessary in the post analyticalphases of cancer genome studies. The "interactive" nature of the paplot-generated reportsenables users to extract much richer information than that obtained from static graphs via most conventional visualization tools.
 
+The worldwide rapid accumulation of massive cancer genome sequencing data presents an unprecedented opportunity for obtain-ing enormous biologically and clinically important information. Several analytical tools have been developed for identifying somat-ic mutations [@koboldt2012varscan; @shiraishi2013empirical], struc-tural variations [@rausch2012delly], and prominent mutation signatures [@shiraishi2015simple] and so on. 
+However, the subse-quent visualization and interpretation often demand us additional laborious efforts such as formatting and organizing the result ta-bles via additional scripting, as well as graphic manipulation via drawing tools, occasionally with multiple trial-and-error parameter adjustments. 
+Additionally, as many existing tools only provide static figures [@skidmore2016genvisr; @yin2012ggbio], it is neces-sary to again repeatedly execute drawing tools with different set-tings even for generating slightly different visualizations (e.g., sorting and filtering by specific criteria). 
+Furthermore, even when some interesting characteristics are observed, we often need to bother to refer back to the raw analytical result tables for identify-ing the samples and the exact values of interest, as these may not be shown in the static figures.
+Here, we present [paplot](https://github.com/Genomon-Project/paplot), a suite of programs to generate various reports for commonly needed tasks in cancer genome analysis. 
+paplot accepts text-based input data and configuration files and generates HTML files with dynamic, interactive data visualizations powered by [D3.js](https://d3js.org). 
+By modifying settings in the configuration files, users can freely change various visual aspects such as colors of objects and contents of pop-up windows appearing after DOM events such as mouse-click and mouse-over.
+
+Interactive graphics contain richer information than static ones, thus facilitating effective and intuitive interpretation of analytical results. 
+Owing to the advancement in information technologies and development of application frameworks for easy generation of interactive graphics, using interactive graphics for academic re-search and auto-generated reports for clinical sequencing will be-come more common. 
+We believe that paplot will be useful for academic research and clinical practice.
+
+![Figures](./img/graphs.png)
+
+**Fig. 1. Example of interactive reports created by paplot.**
+
+ - (A) Quality Control Report.
+ - (B) Chromosomal Aberration Report.
+ - (C) Mutation Matrix Report.
+ - (D) Mutational Signature Report.
+
+## Quality Control Report
+
+Confirming the quality of sequence data (e.g., coverage, insert size and mapping ratio) is a fundamental step in sequence analysis. In many cases, the quality control check is performed by visual in-spection of the barplot to check whether there are some samples with strangely heights. However, as adding the X-axis labels for all the samples is not always possible especially when the sample size in the cohort is large, we have to spend excessive time to retrieve the raw table data to identify the samples with the strange features. By using the Quality Control Report generated by paplot (Figure 1A), one can immediately identify such samples via pop-up con-tents.
+
+## Chromosomal Aberration Report
+
+Chromosomal aberrations play important roles in cancer develop-ment [@mitelman2007impact]. 
[…]
