### 2017-11-30 #457 @arfon
> @arfon, @pjotrp, thank you very much for the suggestion.
I have included detailed description on our software. Please let me know if there is any idea for improving the paper.

@aokad - many thanks for adding some more detail to your submission. At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
