### 2017-12-01 #457 @pjotrp
@aokad, it looks very nice! Can you check the generated PDF - there are some extraneous dashes (word breaks) and you may want to introduce a page break for the figure. If you have any plans for the future with this tooling I would add a short description of that too.

https://github.com/openjournals/joss-papers/blob/joss.00457/joss.00457/10.21105.joss.00457.pdf
