+++ paper/paper.md
@@ -40,14 +40,9 @@ Interactive graphics contain richer information than static ones, thus facilitat
 Owing to the advancement in information technologies and development of application frameworks for easy generation of interactive graphics, using interactive graphics for academic re-search and auto-generated reports for clinical sequencing will be-come more common. 
 We believe that paplot will be useful for academic research and clinical practice.
 
-![Figures](graphs.png)
+---
 
-**Fig. 1. Example of interactive reports created by paplot.**
-
- - (A) Quality Control Report.
- - (B) Chromosomal Aberration Report.
- - (C) Mutation Matrix Report.
- - (D) Mutational Signature Report.
+![Example of interactive reports created by paplot. (A) Quality Control Report. (B) Chromosomal Aberration Report. (C) Mutation Matrix Report. (D) Mutational Signature Report.](graphs.png)
 
 ## Quality Control Report
 
