### 2017-12-04 #457 @aokad
@pjotrp, thank you for your suggestion. I have updated the layout of the figure.
@arfon, please re-generate pdf.
