### 2017-12-04 #457 @arfon
> @arfon, please re-generate pdf.

@aokad - you should be able to do this yourself 😁
