### 2017-12-04 #457 @pjotrp
@aokad if your are happy can you make you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? We can then move forward with accepting the submission.
