### 2017-12-07 #457 @arfon
@aokad - no problem. Can you confirm this version of the paper looks good? https://github.com/openjournals/joss-papers/blob/joss.00457/joss.00457/10.21105.joss.00457.pdf
