### 2017-12-12 #457 @aokad
@arfon I'm sorry to bother you again, but as a result of the reconsideration, I changed the arrangement of sentences.
I'm happy if I can update the paper. The DOI of the modified version is 10.5281/zenodo.1098885.
