### 2017-12-12 #457 @arfon
@aokad - please review the [linked PDF above](https://github.com/openjournals/joss-papers/blob/joss.00457/joss.00457/10.21105.joss.00457.pdf) carefully. I would _really_ like to make this the last round of edits 🙏 😁
